requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
numpy>=1.26.0
docx2txt>=0.8
